Put AndroidLogcatSettingsProvider into separate file
