+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatSettings.cs
@@ -282,67 +282,5 @@ namespace Unity.Android.Logcat
             EditorPrefs.SetString(kSettingsName, data);
         }
     }
-
-    class AndroidLogcatSettingsProvider : SettingsProvider
-    {
-        internal static readonly string kSettingsPath = "Preferences/Analysis/Android Logcat Settings";
-
-        class Styles
-        {
-            public static GUIContent maxMessageCount = new GUIContent("Max Count", "The maximum number of messages.");
-            public static GUIContent font = new GUIContent("Font", "Font used for displaying messages");
-            public static GUIContent fontSize = new GUIContent("Font Size");
-        }
-
-        private AndroidLogcatRuntimeBase m_Runtime;
-        private AndroidLogcatRegexList m_RegexList;
-
-        private AndroidLogcatSettings Settings => m_Runtime.Settings;
-
-        public AndroidLogcatSettingsProvider(string path, SettingsScope scope)
-            : base(path, scope)
-        {
-            m_Runtime = AndroidLogcatManager.instance.Runtime;
-            m_RegexList = new AndroidLogcatRegexList(Settings.StacktraceResolveRegex, m_Runtime);
-        }
-
-        public override void OnGUI(string searchContext)
-        {
-            var settings = Settings;
-            EditorGUILayout.LabelField("Messages", EditorStyles.boldLabel);
-            settings.MaxMessageCount = EditorGUILayout.IntSlider(Styles.maxMessageCount, settings.MaxMessageCount, 1, 100000);
-            settings.MessageFont = (Font)EditorGUILayout.ObjectField(Styles.font, settings.MessageFont, typeof(Font), true);
-            settings.MessageFontSize = EditorGUILayout.IntSlider(Styles.fontSize, settings.MessageFontSize, 5, 25);
-
-            GUILayout.Space(20);
-            EditorGUILayout.LabelField("Message Colors", EditorStyles.boldLabel);
-            foreach (var p in (AndroidLogcat.Priority[])Enum.GetValues(typeof(AndroidLogcat.Priority)))
-            {
-                settings.SetMessageColor(p, EditorGUILayout.ColorField(p.ToString(), settings.GetMessageColor(p)));
-            }
-            GUILayout.Space(20);
-            EditorGUILayout.LabelField("Memory Window", EditorStyles.boldLabel);
-            settings.MemoryRequestIntervalMS = EditorGUILayout.IntField("Request Interval ms", settings.MemoryRequestIntervalMS);
-            GUILayout.Space(20);
-
-            EditorGUILayout.LabelField("Stacktrace Resolve Regex", EditorStyles.boldLabel);
-            m_RegexList.OnGUI();
-
-            GUILayout.Space(20);
-            GUILayout.BeginHorizontal();
-            GUILayout.FlexibleSpace();
-            if (GUILayout.Button("Reset"))
-                settings.Reset();
-            GUILayout.Space(5);
-            GUILayout.EndHorizontal();
-        }
-
-        [SettingsProvider]
-        public static SettingsProvider CreateMyCustomSettingsProvider()
-        {
-            var provider = new AndroidLogcatSettingsProvider(kSettingsPath, SettingsScope.User);
-            return provider;
-        }
-    }
 }
 #endif
